fix: refactor parser and adjust file structure
