+++ index.js
@@ -1,5 +1,5 @@
 const usCities = require('./data/us-cities.json');
-const { parser } = require('./parser');
+const { baseParser } = require('./parsers/baseParser');
 
 'use strict';
 
@@ -10,7 +10,7 @@ const { parser } = require('./parser');
  **/
 function parseAddress(address) {
   // Get the parsed result from the parser
-  const result = parser(address);
+  const result = baseParser(address);
   
   // Add format method to the result object
   result.format = function(formatType) {
